refactor: rename level checker to levels controller
