+++ spec/alephant/logger/support/levels_controller_shared_examples.rb
